chore: Adjusting history, navigation events
