+++ src/Uno.UI/UI/Xaml/Controls/WebView/Core/CoreWebView2DOMContentLoadedEventArgs.cs
@@ -1,5 +1,7 @@
 #nullable enable
 
+using System;
+
 namespace Microsoft.Web.WebView2.Core;
 
 /// <summary>
