+++ src/Uno.UI/UI/Xaml/Controls/WebView/IWebView.cs
@@ -2,4 +2,5 @@
 
 internal interface IWebView
 {
+	bool IsLoaded { get; }
 }
